Lab_Tracing_Application_Insights/requirements.txt
asgiref==3.8.1
azure-ai-projects==1.0.0b6
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.20.0
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b33
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Deprecated==1.2.18
fixedint==0.1.6
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
oauthlib==3.2.2
opentelemetry-api==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-dbapi==0.51b0
opentelemetry-instrumentation-django==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-instrumentation-psycopg2==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-instrumentation-wsgi==0.51b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
portalocker==2.10.1
psutil==5.9.8
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
pywin32==308
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
